client-server model
This is chat system in python that use client-server model and it’s an encrypted chat system.
When the client start chat the server send a random key to the client to encrypt the message
using this key. Then the client sends the encrypting message to the serve and the server decrypt
the message and display it.