how many islands are there in the ocean
use picture of an ocean and you are asked to find how many islands are there in
the ocean. The picture can be visualized as a 2D grid where each cell in the grid represents a pixel in
the picture. If a pixel is part of the water (or sea) then it is represented as 1 in the grid while a land (or
island) pixel is represented as 0.